+++ src/oracledb/__init__.py
@@ -287,6 +287,7 @@ from .driver_mode import is_thin_mode as is_thin_mode
 
 from .utils import (
     enable_thin_mode as enable_thin_mode,
+    from_arrow as from_arrow,
     register_params_hook as register_params_hook,
     register_password_type as register_password_type,
     register_protocol as register_protocol,