+++ src/oracledb/errors.py
@@ -322,6 +322,7 @@ ERR_UNSUPPORTED_PIPELINE_OPERATION = 3028
 ERR_INVALID_NETWORK_NAME = 3029
 ERR_ARROW_UNSUPPORTED_DATA_TYPE = 3030
 ERR_ARROW_UNSUPPORTED_VECTOR_FORMAT = 3031
+ERR_ARROW_UNSUPPORTED_DATA_FORMAT = 3032
 
 # error numbers that result in DatabaseError
 ERR_TNS_ENTRY_NOT_FOUND = 4000
@@ -899,6 +900,10 @@ ERR_MESSAGE_FORMATS = {
     ERR_ARROW_SPARSE_VECTOR_NOT_ALLOWED: (
         "Apache Arrow format does not support sparse vectors with flexible "
         "dimensions"
+    ),
+    ERR_ARROW_UNSUPPORTED_DATA_FORMAT: (
+        'conversion from Arrow format "{schema_format}" to Oracle Database '
+        "is not supported"
     ),
     ERR_ARROW_UNSUPPORTED_DATA_TYPE: (
         "conversion from Oracle Database type {db_type_name} to Apache "