+++ tests/test_8000_dataframe.py
@@ -243,7 +243,7 @@ class TestCase(test_env.BaseTestCase):
         Checks to see if the pyarrow and pandas modules are available.
         """
         if not HAS_INTEROP:
-            self.skipTest("missing pandas or pyarrow modules")
+            self.skipTest("missing numpy, pandas or pyarrow modules")
 
     def __convert_date(self, value):
         """
@@ -395,10 +395,7 @@ class TestCase(test_env.BaseTestCase):
         """
         raw_df = self.__convert_to_df(data)
         raw_data = self.__get_data_from_df(raw_df)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         fetched_data = self.__get_data_from_df(fetched_df)
         self.assertEqual(fetched_data, raw_data)
 
@@ -531,10 +528,7 @@ class TestCase(test_env.BaseTestCase):
         ora_df = self.conn.fetch_df_all(
             "select to_clob('test_8023') from dual"
         )
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         fetched_data = self.__get_data_from_df(fetched_df)
         self.assertEqual(fetched_data, data)
 
@@ -545,10 +539,7 @@ class TestCase(test_env.BaseTestCase):
         ora_df = self.conn.fetch_df_all(
             "select to_blob(utl_raw.cast_to_raw('test_8024')) from dual"
         )
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         fetched_data = self.__get_data_from_df(fetched_df)
         self.assertEqual(fetched_data, data)
 
@@ -559,10 +550,7 @@ class TestCase(test_env.BaseTestCase):
         ora_df = self.conn.fetch_df_all(
             "select utl_raw.cast_to_raw('test_8025') from dual"
         )
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         fetched_data = self.__get_data_from_df(fetched_df)
         self.assertEqual(fetched_data, data)
 
@@ -584,10 +572,7 @@ class TestCase(test_env.BaseTestCase):
             select true
             """
         )
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         fetched_data = self.__get_data_from_df(fetched_df)
         self.assertEqual(fetched_data, data)
 
@@ -626,10 +611,7 @@ class TestCase(test_env.BaseTestCase):
             (None,),
             (None,),
         ]
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         fetched_data = self.__get_data_from_df(fetched_df)
         self.assertEqual(fetched_data, data)
 
@@ -657,12 +639,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 2)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        # number of children for a nested list = 1
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 1)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_vectors_supported()
@@ -682,11 +659,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 2)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 1)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_vectors_supported()
@@ -706,11 +679,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 2)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 1)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_vectors_supported()
@@ -730,11 +699,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 2)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 1)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_vectors_supported()
@@ -757,11 +722,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 3)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 1)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_vectors_supported()
@@ -811,11 +772,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 12)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 1)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_sparse_vectors_supported()
@@ -857,12 +814,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 2)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        # number of children for a struct = 3 (num_dimensions, indices, values)
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 3)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_sparse_vectors_supported()
@@ -904,12 +856,7 @@ class TestCase(test_env.BaseTestCase):
         )
         self.assertEqual(ora_df.num_rows(), 2)
         self.assertEqual(ora_df.num_columns(), 1)
-        fetched_tab = pyarrow.Table.from_arrays(
-            ora_df.column_arrays(), names=ora_df.column_names()
-        )
-        # number of children for a struct = 3 (num_dimensions, indices, values)
-        self.assertEqual(fetched_tab.schema.types[0].num_fields, 3)
-        fetched_df = fetched_tab.to_pandas()
+        fetched_df = pyarrow.table(ora_df).to_pandas()
         self.assertEqual(data, self.__get_data_from_df(fetched_df))
 
     @test_env.skip_unless_vectors_supported()